Add files via upload
This is a very simple front end I created. I wanted to make this more collaborative and have some sort of foundation to go off. Formatting is currently all aligned to the left, and there are buttons for the "questions" that we are responsible for answering. Of course, we can change/edit flexibly as we go along. Just wanted to get something out there for now as we have ~13 hrs left if we want to go through with our project.